backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
langchain>=0.3.19
langchain-community==0.3.18
langchain-openai==0.3.0
langchain-huggingface<1.0.0
faiss-cpu>=1.13.0
pypdf==5.1.0
python-dotenv==1.0.1
pydantic>=2.0
pydantic-settings>=2.0
email-validator>=2.0
sentence-transformers>=2.2.0
torch>=2.0.0
